+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 
 ### Bugfixes
+- Adds a default center position when calculating `panX` and `panY` after double click on an aircraft strip [#207](https://github.com/openscope/openscope/issues/207)